Инкремент 19. Исправил multichecker вынеся проверку использования os.Exit в отдельный файл. Переписал реализацию этой проверки. До этого она работала не верно.
